Added types, spacing, moved rgx to own module
